WIP - added experimental support for pointcloud decoding with wasm
